Consolidate StorePatchBuilder and StoreInternalApi test helpers
- Move StorePatchBuilder from :tart-test to :tart-core so the public
  extension and internal tests share a single implementation
- Rename StorePatchTestSupport to StoreInternalApiTestSupport and add
  dispatchAndWaitForTest / startAndWaitForTest; StoreBaseTest and
  StoreObserverTest drop their local duplicates
- Narrow the @InternalTartApi gate to StoreInternalApi alone; StorePatch,
  PluginPatch, and StorePatchBuilder become plain public APIs
- Add @Volatile to isConfigurationFrozen so writes from lazy
  initialization publish to patch() readers across threads
- Give PluginPatch.Clear explicit equals/hashCode/toString to match the
  data class siblings
