+++ tart-core/src/commonMain/kotlin/io/yumemi/tart/core/StoreImpl.kt
@@ -20,6 +20,7 @@ import kotlinx.coroutines.isActive
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
+import kotlin.concurrent.Volatile
 import kotlin.coroutines.CoroutineContext
 import kotlin.coroutines.EmptyCoroutineContext
 import kotlin.reflect.KClass
@@ -129,6 +130,7 @@ internal abstract class StoreImpl<S : State, A : Action, E : Event> : Store<S, A
 
     private var activeDispatchJob: Job? = null
 
+    @Volatile
     private var isConfigurationFrozen: Boolean = false
 
     private var isInitialized: Boolean = false
