+++ tart-core/src/commonTest/kotlin/io/yumemi/tart/core/StoreBaseTest.kt
@@ -11,7 +11,7 @@ import kotlin.test.assertFalse
 import kotlin.test.assertIs
 import kotlin.test.assertNotNull
 
-@OptIn(kotlinx.coroutines.ExperimentalCoroutinesApi::class, InternalTartApi::class)
+@OptIn(kotlinx.coroutines.ExperimentalCoroutinesApi::class)
 class StoreBaseTest {
 
     private val testDispatcher = UnconfinedTestDispatcher()
@@ -301,12 +301,5 @@ class StoreBaseTest {
 
         assertNotNull(emittedEvent)
         assertEquals(AppEvent.CountUpdated(1), emittedEvent)
-    }
-
-    private suspend fun Store<AppState, AppAction, AppEvent>.dispatchAndWaitForTest(action: AppAction) {
-        @Suppress("UNCHECKED_CAST")
-        val storeInternalApi = this as? StoreInternalApi<AppState, AppAction, AppEvent>
-            ?: error("Expected Tart Store to implement StoreInternalApi")
-        storeInternalApi.dispatchAndWait(action)
     }
 }
