+++ tart-core/src/commonTest/kotlin/io/yumemi/tart/core/StorePatchTest.kt
@@ -12,7 +12,7 @@ import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 import kotlin.test.assertTrue
 
-@OptIn(InternalTartApi::class, ExperimentalCoroutinesApi::class)
+@OptIn(ExperimentalCoroutinesApi::class)
 class StorePatchTest {
 
     data class AppState(val count: Int) : State
